refactor: make repository functions modular and improve query
