--- src/app/@types/migrations.ts
@@ -1,4 +1,0 @@
-import { FailureError } from '@/logic/result';
-
-export type MigrationsErrorCodes = 'UNKNOWN';
-export type MigrationsError = FailureError<MigrationsErrorCodes>;
